apk_verify: [SECURITY FIX] Explicitly specify name of certificate file
This commit adds explicit specification of the certificate file name.
Without this change, a tampered-with APK file could possibly be accepted
if it contains the trusted certificate file, but is actually signed by
another file. (I haven't tested with a such APK file, and I don't know if
Android would accept it, though).